+++ src/Exception/DelegatedCompileException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace BEAR\Package\Exception;
-
-final class DelegatedCompileException extends LogicException
-{
-}