+++ tests/Fake/fake-app/bin/compile.php
@@ -12,7 +12,6 @@ declare(strict_types=1);
  */
 
 use BEAR\Package\Compiler;
-use BEAR\Package\Injector;
 
 require dirname(__DIR__) . '/vendor/autoload.php';
 
@@ -22,4 +21,4 @@ is_file($dotCompile) && require $dotCompile;
 
 $context = $argv[1] ?? 'prod-app';
 
-exit(Compiler::fromInjector(Injector::getInstance('FakeVendor\HelloWorld', $context, dirname(__DIR__)), $context)());
+exit((new Compiler('FakeVendor\HelloWorld', $context, dirname(__DIR__)))());